Some changes

+++ src/main/java/com/assessing/project/model/repository/AdminRepository.java
@@ -3,6 +3,6 @@ package com.assessing.project.model.repository;
 import com.assessing.project.model.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
     Admin findAdminByLogin(String login);
 }
